+++ sql/constraints/fk_music_rating_job_id_job.sql
@@ -1,6 +1,0 @@
-ALTER TABLE music_rating
-    ADD CONSTRAINT fk_music_rating_job_id_job
-        FOREIGN KEY (job_id)
-            REFERENCES job (id)
-            ON DELETE CASCADE
-            ON UPDATE CASCADE;