+++ sql/types/play_result_rank.sql
@@ -1,1 +1,0 @@
-CREATE TYPE play_result_rank AS ENUM ('D', 'C', 'B', 'BB', 'BBB', 'A', 'AA', 'AAA', 'S', 'S+', 'SS', 'SS+', 'SSS', 'SSS+');