+++ sql/types/rating_type.sql
@@ -1,1 +1,0 @@
-CREATE TYPE rating_type AS ENUM ('best', 'recent', 'selection');