+++ sql/types/std_chart_difficulty.sql
@@ -1,1 +1,0 @@
-CREATE TYPE std_chart_difficulty AS ENUM ('basic', 'advanced', 'expert', 'master', 'ultima');