+++ src/index.ts
@@ -1,21 +1,16 @@
 import { chromium } from "playwright";
-import postgres from "postgres";
 
-import { environment } from "./environment.js";
+import { db } from "./db/index.js";
+import { job } from "./db/schema/index.js";
+import { PwPage } from "./playwright.js";
 import { login } from "./steps/1-login.js";
-import { playerData } from "./steps/2-playerdata.js";
+import { scrapePlayerData } from "./steps/2-playerdata.js";
 import { qman } from "./steps/3-qman.js";
 import { genImage } from "./steps/4-genimage.js";
 
-const sql = postgres(environment.DATABASE_URL);
-
-const jobId = (
-  await sql`INSERT INTO job (timestamp) VALUES (NOW()) RETURNING id`
-)[0].id as number;
+const jobId = (await db.insert(job).values({}).returning())[0].id;
 
 console.log(`Created job with ID: ${jobId}`);
-
-const start = performance.now();
 
 const browser = await chromium.launch({
   headless: !process.env.DEBUG,
@@ -23,33 +18,30 @@ const browser = await chromium.launch({
 });
 
 const page = await browser.newPage();
+const pwPage = new PwPage(jobId, page, db);
 
 // * Step 1: Login
-await login(page);
-await page.getByRole("link", { name: "PLAYER DATA" }).click();
-await page.waitForURL("https://chunithm-net-eng.com/mobile/home/playerData");
-
-const step1Time = performance.now();
-console.log(`Step 1 Completed: Took ${Math.round(step1Time - start)}ms`);
+await pwPage.runStep("Step 1: Login", (_, page) => login(page));
+await pwPage.runStep("Prepare for Step 2", async (_, page) => {
+  await page.getByRole("link", { name: "PLAYER DATA" }).click();
+  await page.waitForURL("https://chunithm-net-eng.com/mobile/home/playerData");
+});
 
 // * Step 2: Player Data
-const lastPlayed = await playerData(page, sql, jobId);
-
-const step2Time = performance.now();
-console.log(`Step 2 Completed: Took ${Math.round(step2Time - step1Time)}ms`);
+const lastPlayed = await pwPage.runStep(
+  "Step 2: Player Data",
+  scrapePlayerData,
+);
 
 // * Step 3: Qman
-const tempFileLocation = "./temp.json";
-await qman(page, sql, jobId, lastPlayed, tempFileLocation);
-
-const step3Time = performance.now();
-console.log(`Step 3 Completed: Took ${Math.round(step3Time - step2Time)}ms`);
+const qmanResult = await pwPage.runStep("Step 3: Qman", (jobId, page, db) =>
+  qman(jobId, page, db, lastPlayed),
+);
 
 // * Step 4: Generate Image
-await genImage(page, tempFileLocation);
+await pwPage.runStep("Step 4: Generate Image", (_, page) =>
+  genImage(page, qmanResult),
+);
 
-const step4Time = performance.now();
-console.log(`Step 4 Completed: Took ${Math.round(step4Time - step3Time)}ms`);
-
-await sql.end();
+await db.$client.end();
 await browser.close();