+++ src/steps/3-qman.ts
@@ -1,88 +1,93 @@
 /// <reference types="bun-types" />
 
 import { Page } from "playwright";
-import { Sql } from "postgres";
 
 import {
   calculateRank,
   calculateRating,
   getInternalLevel,
 } from "../data/findRating.js";
+import {
+  chartScore,
+  musicRating,
+  qmanRawTable,
+  RatingType,
+} from "../db/schema/index.js";
 import { environment } from "../environment.js";
-import { PlayResultRank, RatingType, StdChartDifficulty } from "../types.js";
+import { dbType } from "../playwright.js";
 
 import { main } from "./vendor/qman.js";
 import { QmanData, QmanScore, toStdDifficulty } from "./vendor/types.js";
 
 async function uploadMusic(
-  sql: Sql,
-  scores: QmanScore[],
   jobId: number,
+  db: dbType,
+  scores: QmanScore[],
   ratingType: RatingType,
 ) {
-  const payload = [] as Array<{
-    job_id: number;
-    title: string;
-    score: number;
-    difficulty: StdChartDifficulty;
-    rating_type: RatingType;
-    music_order: number;
-    level: number;
-    rank: PlayResultRank;
-    rating: number;
-  }>;
+  const payload = [] as (typeof musicRating.$inferInsert)[];
 
   for (const [index, score] of scores.entries()) {
     const level = await getInternalLevel(
       score.title,
       toStdDifficulty(score.difficulty),
-      sql,
+      db,
       environment.VERSION,
     );
 
     payload.push({
-      job_id: jobId,
+      jobId: jobId,
       title: score.title,
       score: score.score,
       difficulty: toStdDifficulty(score.difficulty),
-      rating_type: ratingType,
-      music_order: index + 1,
-      level: level,
+      ratingType: ratingType,
+      musicOrder: index + 1,
+      level: level.toString(),
       rank: calculateRank(score.score),
-      rating: calculateRating(score.score, level),
+      rating: calculateRating(score.score, level).toFixed(2),
     });
   }
 
-  await sql`INSERT INTO music_rating ${sql(
-    payload,
-    "job_id",
-    "title",
-    "score",
-    "difficulty",
-    "rating_type",
-    "music_order",
-    "level",
-    "rank",
-    "rating",
-  )}`;
+  await db.insert(musicRating).values(payload);
 }
 
 export async function qman(
-  page: Page,
-  sql: Sql,
   jobId: number,
+  page: Page,
+  db: dbType,
   lastPlayed: Date,
-  tempFileLocation: string,
 ) {
   const result = (await page.evaluate(main)) as QmanData;
 
-  await uploadMusic(sql, result.best, jobId, "best");
-  await uploadMusic(sql, result.recent, jobId, "recent");
-  await uploadMusic(sql, result.candidate, jobId, "selection");
+  await uploadMusic(jobId, db, result.best, "best");
+  await uploadMusic(jobId, db, result.recent, "recent");
+  await uploadMusic(jobId, db, result.candidate, "selection");
 
   for (const score of result.score) {
-    await sql`INSERT INTO chart_score (job_id, title, difficulty, score, fc, aj, clear_mark, full_chain, updated_at) VALUES (${jobId}, ${score.title}, ${toStdDifficulty(score.difficulty)}, ${score.score}, ${score.isFullCombo}, ${score.isAllJustice}, ${score.clearMark}, ${score.fullChain}, ${lastPlayed}) ON CONFLICT DO NOTHING`;
+    await db
+      .insert(chartScore)
+      // @ts-expect-error insert on table with default column
+      .values({
+        jobId: jobId,
+        title: score.title,
+        difficulty: toStdDifficulty(score.difficulty),
+        score: score.score,
+        fc: score.isFullCombo,
+        aj: score.isAllJustice,
+        clearMark: score.clearMark,
+        fullChain: score.fullChain,
+        updatedAt: lastPlayed.toISOString(),
+      })
+      .onConflictDoNothing();
   }
 
-  await Bun.write(tempFileLocation, JSON.stringify(result));
+  const resultStr = JSON.stringify(result);
+
+  // @ts-expect-error insert on table with default column
+  await db.insert(qmanRawTable).values({
+    jobId: jobId,
+    raw: resultStr,
+  });
+
+  return resultStr;
 }