+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.0.0] - 2025-03-03
+
+- feat!: migrate to drizzle orm
+- feat: add error handling and send screenshot when failed
+
 ## [2.2.1] - 2025-02-27
 
 - fix: handle case where chart constant is not found